Move new deposits exception to exception package
